Separate the target pixel buffer trait and related types into a separate module
